requirements.txt
streamlit
opencv-python
numpy
pandas
pillow
